continue refactoring #91959
